requirements.txt
opencv-python
pillow
requests